Moved LayoutElementParcelable to adapters.data
