+++ app/src/main/java/com/amaze/filemanager/adapters/data/LayoutElementParcelable.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.ui;
+package com.amaze.filemanager.adapters.data;
 
 import android.os.Parcel;
 import android.os.Parcelable;
